Move front-proxy constants, vars, and loadFrontProxyConfig to config.go
Keep all package variables, constants, and init-time loading logic in
one place, as requested in upstream review. frontproxy.go now only
contains the verifiedFrontProxy request check.
